refactor: reorganize account management components and enhance address handling
