+++ app/myaccount/components/MyReview.js
@@ -1,0 +1,10 @@
+export default function MyReviews() {
+  return (
+    <div className="w-full flex flex-col items-center">
+      <h2 className="text-xl sm:text-2xl font-bold text-center text-blue-900 mb-4">
+        MY REVIEWS
+      </h2>
+      <div className="text-gray-400 text-lg mt-12">No data record.</div>
+    </div>
+  );
+}
